added edit and delete button in comments section

+++ Invite/js/app.js
@@ -1197,8 +1197,10 @@ async function verifyOtp(mobileNumber, otp, callBack) {
             otp
         })
     });
-    if (resp.ok && typeof callBack === 'function') {
-        callBack();
+    if (resp.ok) {
+        if (typeof callBack === 'function') {
+            callBack();
+        }
     } else {
         await showApiErr(resp);
     }
@@ -1313,18 +1315,29 @@ async function addComment(message, parentCommentId = "") {
  * @param {string} message 
  * @param {string} commentId 
  */
-async function editComment(message, commentId) {
-    const resp = await fetch(`${apiUrl}/comments/${commentId}`, {
-        method: "PUT",
-        headers: getHeaders(),
-        body: JSON.stringify({
-            message
-        })
-    })
-    if (resp.ok) {
-        alert('Done');
-    } else {
-        alert(await resp.text());
+
+async function editComment(commentId, message) {
+    const newMessage = prompt('Edit your comment:', message);
+    if (newMessage) {
+        try {
+            const resp =  await fetch(`${apiUrl}/comments/${commentId}`, {
+                method: 'PUT',
+                headers: {
+                    'Content-Type': 'application/json',
+                    'Authorization': `Bearer ${getAccessToken()}`
+                },
+                body: JSON.stringify({ message: newMessage })
+            });
+            if (resp.ok) {
+                alert('Comment edited successfully');
+                renderComment(true);
+            } else {
+                alert(await resp.text());
+            }
+        } catch (error) {
+            console.error('Error editing comment:', error);
+            alert('An error occurred while editing the comment.');
+        }
     }
 }
 
@@ -1333,17 +1346,26 @@ async function editComment(message, commentId) {
  * @param {string} commentId 
  */
 async function deleteComment(commentId) {
-    const resp = await fetch(`${apiUrl}/comments/${commentId}`, {
-        method: "DELETE",
-        headers: getHeaders()
-    })
-    if (resp.ok) {
-        alert('Done');
-    } else {
-        alert(await resp.text());
+    if (confirm('Are you sure you want to delete this comment?')) {
+        try {
+            const resp = await fetch(`${apiUrl}/comments/${commentId}`, {
+                method: 'DELETE',
+                headers: {
+                    'Authorization': `Bearer ${getAccessToken()}`
+                }
+            });
+            if (resp.ok) {
+                alert('Comment deleted successfully');
+                renderComment(true);
+            } else {
+                alert("Cannot delete other messages");
+            }
+        } catch (error) {
+            console.error('Error deleting comment:', error);
+            alert('An error occurred while deleting the comment.');
+        }
     }
 }
-
 /**
  * 
  * @param {number} score 
@@ -1497,23 +1519,33 @@ const renderComment = async (reload = false, filters = {}) => {
         $('#load-cmt').hide();
     } else {
         const commentHtml = comments.map(cmt => {
-            const avatarUrl = `https://ui-avatars.com/api/?name=${encodeURIComponent(cmt.user.name)}&background=random&color=fff&size=40&font-size=0.6`;
+            const avatarUrl = `https://ui-avatars.com/api/?name=${encodeURIComponent(cmt.user.name)}&background=random&color=fff&size=40&font-size=0.6`; 
             return `
-                <div class="mcomment">
-                    <div class="mcomment-header">
-                        <img src="${avatarUrl}" alt="Avatar" class="mavatar">
-                        <div class="mcomment-header-text">
-                            <div class="muser-info">
-                                <span class="musername">${cmt.user.name}</span>
-                                <span class="mdate">${moment(cmt.updatedAt).fromNow()}</span>
-                            </div>
-                            <span class="mrelationType">${cmt.user.relationType}</span>
+            <div class="mcomment">
+                <div class="mcomment-header">
+                    <img src="${avatarUrl}" alt="Avatar" class="mavatar">
+                    <div class="mcomment-header-text">
+                        <div class="muser-info">
+                            <span class="musername">${cmt.user.name}</span>
+                            <span class="mdate">${moment(cmt.updatedAt).fromNow()}</span>
                         </div>
+                        <span class="mrelationType">${cmt.user.relationType}</span>
                     </div>
-                    <div class="mcomment-body">
-                        ${cmt.message}
-                    </div>
+                    ${
+                        cmt.canModify ? `<div class="mcomment-actions">
+                        <button class="edit-btn" style="background-color: #674ea7;" onclick="editComment('${cmt.id}', '${cmt.message}')">
+                            <i class="fa-solid fa-pen-to-square"></i>
+                        </button>
+                        <button class="delete-btn" style="background-color: #f44336;" onclick="deleteComment('${cmt.id}')">
+                            <i class="fa-solid fa-trash"></i>
+                        </button>
+                    </div>` : ''
+                    }
                 </div>
+                <div class="mcomment-body">
+                    ${cmt.message}
+                </div>
+            </div>
             `;
         }).join('');
 
@@ -1563,13 +1595,14 @@ const applyFilters = debounce(() => {
 }, 300);
 
 window.attendeesContent = `
-<h4>Attendees</h4>
+ <h4 style="visibility: hidden;">Attendees</h4>
 <ul class="attendeesList mcomments">
     <div class="no-records">No records found</div>
 </ul>
 `;
 
 window.absenteesContent = `
+ <h4 style="visibility: hidden;">Attendees</h4>
 <ul class="nattendeesList mcomments">
     <div class="no-records">No records found</div>
 </ul>
@@ -1579,7 +1612,7 @@ window.absenteesContent = `
 const renderAttendeesContent = async (reload = false, filters = {}) => {
     $('#attendees-content').html(window.attendeesContent);
     $('#absentees-content').html(window.absenteesContent);
-   
+
     if (cachedAttendees.length === 0 || reload) {
         showLoader('.attendeesList');
         showLoader('.nattendeesList');
@@ -1632,6 +1665,25 @@ const renderAttendeesContent = async (reload = false, filters = {}) => {
         }
     });
 };
+
+function editAttendee(attendeeId, name, relationType, colleagueRef) {
+    const newName = prompt('Edit name:', name);
+    const newRelationType = prompt('Edit relation type:', relationType);
+    const newColleagueRef = prompt('Edit colleague reference:', colleagueRef);
+    if (newName && newRelationType && newColleagueRef) {
+       
+        window.alert("asd");
+        renderAttendeesContent(true);
+    }
+}
+
+function deleteAttendee(attendeeId) {
+    if (confirm('Are you sure you want to delete this attendee?')) {
+        
+        window.alert("ajith");
+        renderAttendeesContent(true);
+    }
+}
 
 function debounce(func, wait) {
     let timeout;
@@ -1744,5 +1796,3 @@ $(document).ready(function(){
         $('#contentBlockAnu').hide();
     });
 });
-
-
